refactor: extract normalizeTechniqueValue and techniqueDisplay to shared utility
