+++ src/lib/components/saved/SavedScreen.svelte
@@ -1,6 +1,7 @@
 <script>
   import { getContext } from "svelte";
   import { anxietyLabel } from "../../anxiety.js";
+  import { normalizeTechniqueValue, techniqueDisplay } from "../../technique-utils.js";
 
   const store = getContext("app");
 
@@ -121,17 +122,6 @@
     if (!iso) return "";
     const d = new Date(iso);
     return d.toLocaleDateString(undefined, { month: "short", day: "numeric", year: "numeric" });
-  }
-
-  function normalizeTechniqueValue(value) {
-    if (!Array.isArray(value)) return String(value || "");
-    return value.filter((technique) => technique && technique !== "none").slice().sort().join(",");
-  }
-
-  function techniqueDisplay(value) {
-    if (!Array.isArray(value)) return value || null;
-    const normalized = value.filter((technique) => technique && technique !== "none").slice().sort();
-    return normalized.length ? normalized.join(", ") : null;
   }
 
   // Transition notes — same logic as SetlistSongCard
